deployment/requirements.txt
Flask==3.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
tifffile>=2023.7.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Werkzeug>=3.0.0
pandas>=2.0.0
seaborn>=0.12.0
torchsummary>=1.5.1
